+++ app/src/main/java/sample/kingja/loadsir/MainActivity.java
@@ -2,11 +2,12 @@ package sample.kingja.loadsir;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
+import androidx.appcompat.app.AppCompatActivity;
 import sample.kingja.loadsir.target.AnimateActivity;
 import sample.kingja.loadsir.target.BestPracticesActivity;
+import sample.kingja.loadsir.target.ConstraintLayoutActivity;
 import sample.kingja.loadsir.target.DefaultCallbackActivity;
 import sample.kingja.loadsir.target.KeepTitleActivity;
 import sample.kingja.loadsir.target.KeepTitleFragmentActivity;
@@ -79,5 +80,9 @@ public class MainActivity extends AppCompatActivity {
 
     public void keepTitleInActivity(View view) {
         startActivity(new Intent(this, KeepTitleActivity.class));
+    }
+
+    public void inConstraintLayoutActivity(View view) {
+        startActivity(new Intent(this, ConstraintLayoutActivity.class));
     }
 }