+++ app/src/main/java/sample/kingja/loadsir/base/BaseTitleActivity.java
@@ -1,8 +1,6 @@
 package sample.kingja.loadsir.base;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.FrameLayout;
 import android.widget.LinearLayout;
@@ -12,6 +10,8 @@ import com.kingja.loadsir.callback.Callback;
 import com.kingja.loadsir.core.LoadService;
 import com.kingja.loadsir.core.LoadSir;
 
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 import butterknife.ButterKnife;
 import sample.kingja.loadsir.R;
 
@@ -37,28 +37,18 @@ public abstract class BaseTitleActivity extends AppCompatActivity {
 
 
     private void addContent() {
-        FrameLayout flContent = (FrameLayout) rootView.findViewById(R.id.fl_content);
-        TextView tvTitleTitle = (TextView) rootView.findViewById(R.id.tv_title_title);
-        LinearLayout llTitleBack = (LinearLayout) rootView.findViewById(R.id.ll_title_back);
+        FrameLayout flContent = rootView.findViewById(R.id.fl_content);
+        TextView tvTitleTitle =  rootView.findViewById(R.id.tv_title_title);
+        LinearLayout llTitleBack =  rootView.findViewById(R.id.ll_title_back);
         tvTitleTitle.setText(getContentTitle() == null ? "" : getContentTitle());
-        llTitleBack.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                backClick();
-            }
-        });
+        llTitleBack.setOnClickListener(v -> backClick());
         View content = View.inflate(this, getContentView(), null);
         if (content != null) {
             FrameLayout.LayoutParams params = new FrameLayout.LayoutParams(FrameLayout.LayoutParams.MATCH_PARENT,
                     FrameLayout.LayoutParams.MATCH_PARENT);
             flContent.addView(content, params);
             ButterKnife.bind(this, rootView);
-            mBaseLoadService = LoadSir.getDefault().register(content, new Callback.OnReloadListener() {
-                @Override
-                public void onReload(View v) {
-                    onNetReload(v);
-                }
-            });
+            mBaseLoadService = LoadSir.getDefault().register(content, (Callback.OnReloadListener) this::onNetReload);
         }
     }