+++ app/src/main/java/sample/kingja/loadsir/target/ConstraintLayoutActivity.java
@@ -3,6 +3,7 @@ package sample.kingja.loadsir.target;
 import android.os.Bundle;
 import android.os.SystemClock;
 import android.view.View;
+import android.widget.TextView;
 
 import com.kingja.loadsir.callback.Callback;
 import com.kingja.loadsir.core.LoadService;
@@ -12,8 +13,8 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import sample.kingja.loadsir.PostUtil;
 import sample.kingja.loadsir.R;
-import sample.kingja.loadsir.callback.AnimateCallback;
 import sample.kingja.loadsir.callback.EmptyCallback;
+import sample.kingja.loadsir.callback.LoadingCallback;
 
 
 /**
@@ -23,32 +24,30 @@ import sample.kingja.loadsir.callback.EmptyCallback;
  * Email:kingjavip@gmail.com
  */
 
-public class AnimateActivity extends AppCompatActivity {
-
+public class ConstraintLayoutActivity extends AppCompatActivity {
 
     private LoadService loadService;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_content);
-        // Your can change the callback on sub thread directly.
-        LoadSir loadSir = new LoadSir.Builder()
-                .addCallback(new EmptyCallback())
-                .addCallback(new AnimateCallback())
-                .setDefaultCallback(AnimateCallback.class)
-                .build();
-        loadService = loadSir.register(this, new Callback.OnReloadListener() {
+        setContentView(R.layout.activity_constraintlayout);
+        initLoadSir();
+    }
+
+    private void initLoadSir() {
+        TextView tv_center = findViewById(R.id.tv_center);
+        loadService = LoadSir.getDefault().register(tv_center, new Callback.OnReloadListener() {
             @Override
             public void onReload(View v) {
                 // Your can change the status out of Main thread.
                 new Thread(new Runnable() {
                     @Override
                     public void run() {
-                        loadService.showCallback(AnimateCallback.class);
+                        loadService.showCallback(LoadingCallback.class);
                         //do retry logic...
                         SystemClock.sleep(500);
-                        //callback on sub thread
+                        //callback
                         loadService.showSuccess();
                     }
                 }).start();