+++ app/src/main/java/sample/kingja/loadsir/target/FragmentSingleActivity.java
@@ -1,8 +1,8 @@
 package sample.kingja.loadsir.target;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 
+import androidx.appcompat.app.AppCompatActivity;
 import sample.kingja.loadsir.R;
 
 /**