+++ app/src/main/java/sample/kingja/loadsir/target/NormalFragment.java
@@ -1,22 +1,25 @@
 package sample.kingja.loadsir.target;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+
+import android.os.SystemClock;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import com.kingja.loadsir.callback.Callback;
-
-import sample.kingja.loadsir.callback.CustomCallback;
-import sample.kingja.loadsir.callback.LoadingCallback;
-
 import com.kingja.loadsir.core.LoadService;
 import com.kingja.loadsir.core.LoadSir;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import butterknife.ButterKnife;
 import sample.kingja.loadsir.PostUtil;
 import sample.kingja.loadsir.R;
+import sample.kingja.loadsir.callback.CustomCallback;
+import sample.kingja.loadsir.callback.ErrorCallback;
+import sample.kingja.loadsir.callback.LoadingCallback;
 
 /**
  * Description:TODO
@@ -27,26 +30,41 @@ import sample.kingja.loadsir.R;
 public class NormalFragment extends Fragment {
 
     private LoadService loadService;
-    private View rootView;
 
     @Nullable
     @Override
-    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle
-            savedInstanceState) {
-        rootView = View.inflate(getActivity(), R.layout.fragment_a_content, null);
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
+                             @Nullable Bundle savedInstanceState) {
         LoadSir loadSir = new LoadSir.Builder()
                 .addCallback(new CustomCallback())
                 .addCallback(new LoadingCallback())
+                .addCallback(new ErrorCallback())
                 .setDefaultCallback(LoadingCallback.class)
                 .build();
-        loadService = loadSir.register(rootView);
+        View rootView = inflater.inflate(R.layout.fragment_a_content, container, false);
+        ButterKnife.bind(this, rootView);
+        loadService = loadSir.register(rootView, new Callback.OnReloadListener() {
+            @Override
+            public void onReload(View v) {
+                // Your can change the status out of Main thread.
+                new Thread(new Runnable() {
+                    @Override
+                    public void run() {
+                        loadService.showCallback(LoadingCallback.class);
+                        //do retry logic...
+                        SystemClock.sleep(500);
+                        //callback
+                        loadService.showCallback(CustomCallback.class);
+                    }
+                }).start();
+            }
+        });
         return loadService.getLoadLayout();
     }
 
     @Override
-    public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        PostUtil.postCallbackDelayed(loadService, CustomCallback.class);
+        PostUtil.postCallbackDelayed(loadService, ErrorCallback.class);
     }
-
 }