+++ app/src/main/java/sample/kingja/loadsir/target/ViewTargetActivity.java
@@ -1,20 +1,18 @@
 package sample.kingja.loadsir.target;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.view.View;
 import android.widget.ImageView;
-
-import sample.kingja.loadsir.callback.LoadingCallback;
 
 import com.kingja.loadsir.callback.Callback;
 import com.kingja.loadsir.core.LoadService;
 import com.kingja.loadsir.core.LoadSir;
 
-import sample.kingja.loadsir.R;
-import sample.kingja.loadsir.callback.TimeoutCallback;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 import sample.kingja.loadsir.PostUtil;
+import sample.kingja.loadsir.R;
+import sample.kingja.loadsir.callback.LoadingCallback;
+import sample.kingja.loadsir.callback.TimeoutCallback;
 
 
 /**
@@ -32,21 +30,18 @@ public class ViewTargetActivity extends AppCompatActivity {
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_view);
-        ImageView imageView = (ImageView) findViewById(R.id.iv_img);
+        ImageView imageView = findViewById(R.id.iv_img);
         LoadSir loadSir = new LoadSir.Builder()
                 .addCallback(new TimeoutCallback())
                 .addCallback(new LoadingCallback())
                 .setDefaultCallback(LoadingCallback.class)
                 .build();
-        loadService = loadSir.register(imageView, new Callback.OnReloadListener() {
-            @Override
-            public void onReload(View v) {
-                loadService.showCallback(LoadingCallback.class);
-                //do retry logic...
+        loadService = loadSir.register(imageView, (Callback.OnReloadListener) v -> {
+            loadService.showCallback(LoadingCallback.class);
+            //do retry logic...
 
-                //callback
-                PostUtil.postSuccessDelayed(loadService);
-            }
+            //callback
+            PostUtil.postSuccessDelayed(loadService);
         });
         PostUtil.postCallbackDelayed(loadService, TimeoutCallback.class);
     }